radardistlimit server side variable
Former-commit-id: 877890695fc9b2fe95bbc92398f1f22372287e4d
Former-commit-id: e71e98ef107fdeba63ed10aabd5148385ba2f234

+++ src/game/game.h
@@ -1573,7 +1573,7 @@ namespace hud
     extern void drawpointertex(const char *tex, int x, int y, int s, float r = 1, float g = 1, float b = 1, float fade = 1);
     extern void drawpointer(int w, int h, int index);
     extern int numteamkills();
-    extern float radarrange();
+    extern int radarrange();
     extern void drawblip(const char *tex, float area, int w, int h, float s, float blend, int style, vec &pos, const vec &colour = vec(1, 1, 1), const char *font = "reduced", const char *text = NULL, ...);
     extern int drawprogress(int x, int y, float start, float length, float size, bool left, float r = 1, float g = 1, float b = 1, float fade = 1, float skew = 1, const char *font = NULL, const char *text = NULL, ...);
     extern int drawitembar(int x, int y, float size, bool left, float r = 1, float g = 1, float b = 1, float fade = 1, float skew = 1, float amt = 1, int type = 0);
